Move category filter to filter folder
